refactor : Artie 토큰을 도메인 레이어로 이동
