+++ src/main/java/com/yapp/artie/global/authentication/JwtFilter.java
@@ -1,7 +1,7 @@
 package com.yapp.artie.global.authentication;
 
-import com.yapp.artie.domain.user.adapter.out.authentication.ArtieToken;
 import com.yapp.artie.domain.user.application.port.out.JwtService;
+import com.yapp.artie.domain.user.domain.ArtieToken;
 import java.io.IOException;
 import java.util.NoSuchElementException;
 import javax.servlet.FilterChain;
